fix(Winning): WinningStatistical이 테스트케이스를 통과하도록 수정
winningPrice가 Integer 범위를 넘어갈 수 있으므로 overflow 문제가 발생함. 따라서 Long으로 변경.

+++ src/main/java/lotto/winning/WinningStatistical.java
@@ -6,7 +6,7 @@ import java.util.Map;
 public class WinningStatistical {
     private final Map<WinningResult, Integer> winningResultCount = new HashMap<>();
     private Integer boughtPrice = 0;
-    private Integer winningPrice = 0;
+    private Long winningPrice = 0L;
 
     public void addWinningResult(WinningResult winningResult) {
         Integer prevWinningResultCount = winningResultCount.getOrDefault(winningResult, 0);
